requirements.txt
aiohttp
pandas
beautifulsoup4
pydantic
pyyaml
pyarrow
